Initial button. Changed utils directory.
utils.py has been moved to a utils directory. A buttons.py file, containing a
button class with a test button, has been created there as well. The
buildbutton command can be run to have the bot to display this button, which if pressed will print hello to the terminal.