FAQ MVP
Bot listens for messages with `?` in `faq`-enabled channels and responds
with the appropriate question and answer in the sample csv if a keyword is
detected.
Next steps would be experimenting with NLP so as to not have hard-coded keywords
and determining which python library for keyword extraction would be ideal.